[cheri] Use a capability type for various offsets
The offset parameter in `onTransferPixelsTo` and `onTransferPixelsFrom`§
is used as either an offset or a pointer depending on context.
Therefore, this must be a type that can preserve capability provenance.

The offset parameter in `GrGLAttribArrayState::set` stores a pointer
value and therefore must be a capability preserving type.

The offset parameter in `GrOpsRenderPass::onDrawIndirect` stores a
pointer value and therefore must be a capability preserving type.

(cherry picked from commit 1fbafb86f30534b047b5c9d348f01ccac4c6588c)